+++ test/functional/utils/nodeStarter.spec.ts
@@ -13,8 +13,8 @@
  */
 import { createServer } from 'net';
 
-import { execAsync, isPortAvailable } from '../../../src/utils/nodeStarter';
-import { startServer, closeServer } from '../../utils';
+import { execAsync, isPortAvailable } from '../../../src/utils/node';
+import { startServer, closeServer } from '../../utils/server';
 
 describe('execAsync', () => {
 	it('Should execute ls command succesfully', async () => {
@@ -24,21 +24,25 @@ describe('execAsync', () => {
 });
 
 describe('isPortAvailable', () => {
-	it('Should execute ls command succesfully', async () => {
-		const port = 9901;
-		try {
-			const isPortAvailableBefore = await isPortAvailable(port);
-			expect(isPortAvailableBefore).toBe(true);
+	it('Should return port availablity properly', async () => {
+		// Keep trying untill a free port is found for testing
+		while (true) {
+			const port = Math.round(Math.random() * 65536);
+			try {
+				const isPortAvailableBefore = await isPortAvailable(port);
+				expect(isPortAvailableBefore).toBe(true);
 
-			const server = createServer();
-			await startServer(server, port);
+				const server = createServer();
+				await startServer(server, port);
 
-			const isPortAvailableNow = await isPortAvailable(port);
-			expect(isPortAvailableNow).toBe(false);
+				const isPortAvailableNow = await isPortAvailable(port);
+				expect(isPortAvailableNow).toBe(false);
 
-			await closeServer(server);
-		} catch (err) {
-			// The port is not available in the host machine
+				await closeServer(server);
+				break;
+			} catch (err) {
+				// The port is not available on the host machine
+			}
 		}
 	});
 });
