+++ test/unit/utils/nodeStarter.spec.ts
@@ -11,16 +11,16 @@
  *
  * Removal or modification of this copyright notice is prohibited.
  */
-import { isLiskCoreV3Running } from '../../../src/utils/nodeStarter';
+import { isAppVersion3 } from '../../../src/utils/node';
 
-describe('isLiskCoreV3Running', () => {
+describe('isAppVersion3', () => {
 	it('Should return true for a version of 3.*.*', async () => {
-		const response = await isLiskCoreV3Running('3.1.1');
+		const response = await isAppVersion3('3.1.1');
 		expect(response).toBe(true);
 	});
 
 	it('Should return false for a version that does not start with 3', async () => {
-		const response = await isLiskCoreV3Running('4.1.1');
+		const response = await isAppVersion3('4.1.1');
 		expect(response).toBe(false);
 	});
 });
